backend/requirements.txt
django
psycopg[binary]
djangorestframework
django-cors-headers